pyproject.toml
[tool.poetry]
name = "cdk-pipeline-sagemaker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "UNLICENSED"
readme = "README.md"
packages = [{include = "cdk_pipeline_sagemaker"}]

[tool.poetry.dependencies]
python = "^3.9"
aws-cdk-lib = "2.50.0"
constructs = "^10.1.149"
yamldataclassconfig = "^1.5.0"
boto3 = "^1.26.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
